requirements.txt
matplotlib==3.8.2
Pillow==10.2.0
torch==2.1.2
torchvision==0.16.2

aiogram~=3.3.0